+++ CHANGELOG.md
@@ -15,6 +15,9 @@
   `sourceMapContents`, `sourceMapEmbed`, and `sourceMapRoot` options to
   `render()` and `renderSync()`.
 
+* Fix a bug where passing a relative path to `render()` or `renderSync()` would
+  cause relative imports to break.
+
 ## 1.2.1
 
 * Always emit units in compressed mode for `0` dimensions other than lengths and